Sprite & crash fixes.

+++ DelayedMessageRelay.h
@@ -8,7 +8,7 @@ public:
     bool hasMessage = false, hexMode = false, accurateFloats = false;
     wxString popUpMessage, popUpTitle;
     wxWindow *focusTarget = 0;
-    int unSaved = 0, loadedFileId = 0;
+    size_t unSaved = 0, loadedFileId = 0;
 
     void post(const wxString &message, const wxString &title, wxWindow *target);
 };
